pyproject.toml
[tool.poetry]
name = "crc-bot"
version = "1.3"
description = "Financial Telegram bot with a currency converter and a built-in calculator."
authors = ["Vladislav Andronov <68904879+AndrVLDZ@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
python-dotenv = "1.0.1"
cexprtk = "0.4.1"
aiosqlite = "0.19.0"
aiohttp = "3.8.5"
aiogram = "3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src/crc-bot"
